VitalSignType + VitalSign GET POST PUT DELETE

+++ resq_backend/src/main/java/at/resq/resq_backend/accidentPatient/vitalSign/type/VitalSignTypeRepository.java
@@ -1,7 +1,11 @@
 package at.resq.resq_backend.accidentPatient.vitalSign.type;
 
 
+import at.resq.resq_backend.accidentPatient.vitalSign.VitalSign;
 import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+import java.util.Optional;
 
 /**
  * Project: resQ-app-v1
@@ -10,5 +14,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * Time: 13:52
  */
 
-public interface VitalSignTypeRepository extends JpaRepository<VitalSignType, Integer> {
+public interface VitalSignTypeRepository extends JpaRepository<VitalSignType, Long> {
 }
